+++ src/index.ts
@@ -5,8 +5,7 @@ import { evaluate as doEvaluate } from './evaluator';
 import { readPrompts } from './prompts';
 import providers, { loadApiProvider } from './providers';
 import { loadApiProviders } from './providers';
-import { extractEntities } from './redteam/extraction/entities';
-import { extractSystemPurpose } from './redteam/extraction/purpose';
+import { extractEntities, extractSystemPurpose } from './redteam/extraction';
 import { GRADERS } from './redteam/graders';
 import { Plugins } from './redteam/plugins';
 import { Strategies } from './redteam/strategies';
