+++ src/redteam/index.ts
@@ -10,8 +10,7 @@ import type { TestCaseWithPlugin } from '../types';
 import { maybeLoadFromExternalFile } from '../util';
 import { extractVariablesFromTemplates } from '../util/templates';
 import { HARM_PLUGINS, PII_PLUGINS, ALIASED_PLUGIN_MAPPINGS } from './constants';
-import { extractEntities } from './extraction/entities';
-import { extractSystemPurpose } from './extraction/purpose';
+import { extractEntities, extractSystemPurpose } from './extraction';
 import { Plugins } from './plugins';
 import { CustomPlugin } from './plugins/custom';
 import { loadRedteamProvider } from './providers/shared';
