+++ test/redteam/extraction/index.test.ts
@@ -1,5 +1,5 @@
 import { fetchWithCache } from '../../../src/cache';
-import { extractEntities } from '../../../src/redteam/extraction/entities';
+import { extractEntities, extractSystemPurpose } from '../../../src/redteam/extraction';
 
 jest.mock('../../../src/logger', () => ({
   error: jest.fn(),
@@ -61,5 +61,39 @@ describe('Entities Extractor', () => {
     const result = await extractEntities(['prompt']);
 
     expect(result).toEqual([]);
+  });
+});
+
+describe('System Purpose Extractor', () => {
+  let originalEnv: NodeJS.ProcessEnv;
+
+  beforeEach(() => {
+    originalEnv = process.env;
+    process.env = { ...originalEnv };
+    jest.clearAllMocks();
+    jest.clearAllMocks();
+  });
+
+  afterEach(() => {
+    process.env = originalEnv;
+  });
+
+  it('should use remote generation', async () => {
+    jest.mocked(fetchWithCache).mockResolvedValue({
+      data: { task: 'purpose', result: 'Remote extracted purpose' },
+      cached: false,
+    });
+
+    const result = await extractSystemPurpose(['prompt1', 'prompt2']);
+    expect(result).toBe('Remote extracted purpose');
+    expect(fetchWithCache).toHaveBeenCalledWith(
+      'https://api.promptfoo.dev/v1/generate',
+      expect.objectContaining({
+        method: 'POST',
+        body: JSON.stringify({ task: 'purpose', prompts: ['prompt1', 'prompt2'] }),
+      }),
+      expect.any(Number),
+      'json',
+    );
   });
 });
